fixed all pull request issues
