+++ backend/package.json
@@ -2,10 +2,6 @@
   "dependencies": {
     "@apollo/server": "^4.11.0",
     "@prisma/client": "^5.20.0",
-    "@types/bcrypt": "^5.0.2",
-    "@types/passport": "^1.0.16",
-    "@types/passport-jwt": "^4.0.1",
-    "@types/passport-local": "^1.0.38",
     "bcrypt": "^5.1.1",
     "dotenv": "^16.4.5",
     "graphql": "^16.9.0",
@@ -16,6 +12,10 @@
   },
   "devDependencies": {
     "@types/node": "^22.7.5",
+    "@types/bcrypt": "^5.0.2",
+    "@types/passport": "^1.0.16",
+    "@types/passport-jwt": "^4.0.1",
+    "@types/passport-local": "^1.0.38",
     "prisma": "^5.20.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3"
