--- backend/src/index.js
@@ -1,83 +1,0 @@
-import { ApolloServer } from '@apollo/server'; // preserve-line
-import { startStandaloneServer } from '@apollo/server/standalone'; // preserve-line
-import { PrismaClient } from '@prisma/client';
-const prisma = new PrismaClient();
-// A schema is a collection of type definitions (hence "typeDefs")
-// that together define the "shape" of queries that are executed against
-// your data.
-const typeDefs = `#graphql
-  # Comments in GraphQL strings (such as this one) start with the hash (#) symbol.
-
-  # This "Book" type defines the queryable fields for every book in our data source.
-  type Book {
-    title: String
-    author: String
-  }
-
-  type User {
-    email: String
-    name: String?
-    phone_number: String
-    address: String
-    post_code: String
-    municipality: String
-    username: String
-    password: String
-  }
-
-  type Mutation {
-    makeUser (
-    email: String!
-    name: String!
-    phone_number: String!
-    address: String!
-    municipality: String!
-    username: String!
-    password: String!
-  ): User
-  }
-
-  # The "Query" type is special: it lists all of the available queries that
-  # clients can execute, along with the return type for each. In this
-  # case, the "books" query returns an array of zero or more Books (defined above).
-  type Query {
-    books: [Book]
-  }
-`;
-const books = [
-    {
-        title: 'The Awakening',
-        author: 'Kate Chopin',
-    },
-    {
-        title: 'City of Glass',
-        author: 'Paul Auster',
-    },
-];
-const resolvers = {
-    Query: {
-        books: () => books,
-        findUser: () => {
-        }
-    },
-    Mutation: {
-        makeUser: (root, args) => {
-            const user = { ...args };
-            prisma.user.create(user);
-        }
-    }
-};
-// The ApolloServer constructor requires two parameters: your schema
-// definition and your set of resolvers.
-const server = new ApolloServer({
-    typeDefs,
-    resolvers,
-});
-// Passing an ApolloServer instance to the `startStandaloneServer` function:
-//  1. creates an Express app
-//  2. installs your ApolloServer instance as middleware
-//  3. prepares your app to handle incoming requests
-const { url } = await startStandaloneServer(server, {
-    listen: { port: 4000 },
-});
-console.log(`🚀  Server ready at: ${url}`);
