+++ backend/src/utils/passportSetup.ts
@@ -1,21 +1,14 @@
 import passport from 'passport';
 import { Strategy as LocalStrategy } from 'passport-local';
-
-
-// Prisma
 import { PrismaClient } from '@prisma/client'
-
 const prisma = new PrismaClient() 
-
-// JWT SETUP
 import bcrypt from 'bcryptjs';
 import { Strategy as JwtStrategy, ExtractJwt } from 'passport-jwt';
 
-//const User = require('../models/user');
 
 const JwtStrategyConfiguration = new JwtStrategy({ jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(), secretOrKey: process.env.SECRET }, (async (jwtPayload, done) => {
   try {
-    const user = await prisma.user.findOne({ username: jwtPayload.username });
+    const user = await prisma.user.findUnique({ where: {email: jwtPayload.email } });
     
     if (user) {
       return done(null, user);
@@ -25,7 +18,6 @@ const JwtStrategyConfiguration = new JwtStrategy({ jwtFromRequest: ExtractJwt.fr
     return done(err, false);
   }
 }));
-// JWT SETUP DONE
-// module.exports = JwtStrategyConfiguration;
+
 
 export default JwtStrategyConfiguration
